services/auth/pyproject.toml
[tool.poetry]
name = "auth-api"
version = "0.1.0"
description = "auth api for school management system"
authors = ["a.espahbodi <dev.amir.espahbodi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
environs = {extras = ["django"], version = "^11.0.0"}
psycopg = {extras = ["binary"], version = "^3.1.19"}
django = "^5.0.6"
djangorestframework = "^3.15.1"
django-phonenumber-field = "^7.3.0"
django-phonenumbers = "^1.0.1"
user-agents = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
